requirements.txt
dvc==3.48.0
mlflow==2.11.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
PyYAML==6.0.1
